agenda-deportiva/requirements.txt
requests
beautifulsoup4
pandas
supabase
pytz
python-dotenv
google-generativeai